// app/src/api/jInshanBodyShopApi.schemas.ts
export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const;

export interface PersonalOrder {
  orderId?: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  contactNumber: string | null;
  /** @nullable */
  orderCountry: string | null;
  /** @nullable */
  orderState: string | null;
  /** @nullable */
  orderAddress: string | null;
  /** @nullable */
  orderCity: string | null;
  /** @nullable */
  orderPostalCode: string | null;
  /** @nullable */
  shippingCountry: string | null;
  /** @nullable */
  shippingState: string | null;
  /** @nullable */
  shippingAddress: string | null;
  /** @nullable */
  shippingCity: string | null;
  /** @nullable */
  shippingPostalCode: string | null;
  /** @nullable */
  shippingName: string | null;
  /** @nullable */
  shippingSurname: string | null;
  /** @nullable */
  shippingOption: string | null;
  createdAt?: string;
  status?: OrderStatus;
  number?: number;
  orderValue: number;
  /** @nullable */
  orderName: string | null;
  /** @nullable */
  orderSurname: string | null;
}